Move `is_optional_of` to `util` module
